add complete integration test. WIP
At the moment it doesn't work right because it has a different behaviour than
in a real execution
