+++ src/test/java/uk/ac/ebi/eva/test/configuration/CommonConfiguration.java
