+++ src/test/java/uk/ac/ebi/eva/test/configuration/DatabaseInitializationConfiguration.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.ac.ebi.eva.pipeline.configuration;
+package uk.ac.ebi.eva.test.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
