+++ src/test/java/uk/ac/ebi/eva/test/configuration/GenotypedVcfWorkflowConfiguration.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *          http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.ac.ebi.eva.pipeline.configuration;
+package uk.ac.ebi.eva.test.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -22,11 +22,11 @@ import org.springframework.context.annotation.PropertySource;
 import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
 
 /**
- * Configuration init for AnnotationJob tests
+ * Configuration init for GenotypedVcfJob, from a workflow perspective
  */
 @Configuration
-@PropertySource({"annotation-job.properties"})
-public class AnnotationJobConfiguration {
+@PropertySource({"genotyped-vcf-workflow.properties"})
+public class GenotypedVcfWorkflowConfiguration {
 
     @Bean
     private static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer() {
