+++ src/test/resources/application-integrationTest.properties
@@ -1,0 +1,50 @@
+# JOB
+spring.batch.job.names=load-genotyped-vcf
+spring.profiles.active=test,mongo
+
+# SUBMISSION FIELDS
+input.vcf=target/test-classes/small20.vcf.gz
+input.vcf.id=5
+input.vcf.aggregation=NONE
+
+input.study.name=small vcf
+input.study.id=small_load_2
+input.study.type=COLLECTION
+
+input.pedigree=
+input.gtf=/path/to/gtf/sample.gtf.gz
+input.fasta=/path/to/homo_sapiens/sequence.fa
+
+output.dir=/tmp
+output.dir.annotation=/tmp
+output.dir.statistics=/tmp
+
+db.name=test_full_integration
+#db.collections.variants.name=variants_name_app
+#db.collections.files.name=files_name_app
+
+# EXTERNAL APPLICATIONS
+app.opencga.path=/opt/opencga
+
+app.vep.path=/path/to/vep/executable
+app.vep.num-forks=4
+app.vep.cache.path=/path/to/vep/cache_folder
+app.vep.cache.version=82
+app.vep.cache.species=homo_sapiens
+
+# MONGO DATABASE
+config.db.read-preference=secondary
+db.collections.features.name=features
+db.collections.stats.name=populationStatistics
+
+# STEPS MANAGEMENT
+## Job repository database
+
+## Skip steps
+statistics.skip=false
+annotation.skip=true
+
+## Repeat steps
+## true: The already COMPLETED steps will be rerun. This is restarting the job from the beginning
+## false(default): if the job was aborted and is relaunched, COMPLETEd steps will NOT be done again
+config.restartability.allow=false
